feat(keyboard): extract SettingsShortcutCapture primitive
- Extract KeyRecorder component from KeyboardShortcutsTab as SettingsShortcutCapture
- Add comprehensive tests covering all capture scenarios
- Update KeyboardShortcutsTab to import from new location
- Export component and props type from KeyboardShortcuts barrel

No visual or behavioral changes - pure extraction for reuse.
