+++ src/components/KeyboardShortcuts/index.ts
@@ -1,1 +1,3 @@
 export { ShortcutReferenceDialog } from "./ShortcutReferenceDialog";
+export { SettingsShortcutCapture } from "./SettingsShortcutCapture";
+export type { SettingsShortcutCaptureProps } from "./SettingsShortcutCapture";
